+++ rollup.config.ts
@@ -12,15 +12,10 @@ const treeshake = {
   unknownGlobalSideEffects: false,
 } satisfies RollupOptions["treeshake"];
 
-const classic = {
+const classicCJS = {
   input: "src/classic.ts",
 
   output: [
-    {
-      file: pkg.exports["."].import,
-      format: "esm",
-      sourcemap: false,
-    },
     {
       file: pkg.exports["."].require,
       format: "cjs",
@@ -32,7 +27,18 @@ const classic = {
     rollupPluginAutoExternal(),
     rollupPluginTs({
       transpileOnly: true,
-      tsconfig: "tsconfig.build.json",
+      tsconfig: {
+        fileName: "tsconfig.build.json",
+        hook: (resolvedConfig) => ({
+          ...resolvedConfig,
+          paths: {
+            ...resolvedConfig.paths,
+            "#eslint-plugin-functional/conditional-imports/*": [
+              "src/utils/conditional-imports/cjs/*",
+            ],
+          },
+        }),
+      },
     }),
     rollupPluginDeassert({
       include: ["**/*.{js,ts}"],
@@ -42,15 +48,46 @@ const classic = {
   treeshake,
 } satisfies RollupOptions;
 
-const flat = {
-  input: "src/flat.ts",
+const classicESM = {
+  input: "src/classic.ts",
 
   output: [
     {
-      file: pkg.exports["./flat"].import,
+      file: pkg.exports["."].import,
       format: "esm",
       sourcemap: false,
     },
+  ],
+
+  plugins: [
+    rollupPluginAutoExternal(),
+    rollupPluginTs({
+      transpileOnly: true,
+      tsconfig: {
+        fileName: "tsconfig.build.json",
+        hook: (resolvedConfig) => ({
+          ...resolvedConfig,
+          paths: {
+            ...resolvedConfig.paths,
+            "#eslint-plugin-functional/conditional-imports/*": [
+              "src/utils/conditional-imports/esm/*",
+            ],
+          },
+        }),
+      },
+    }),
+    rollupPluginDeassert({
+      include: ["**/*.{js,ts}"],
+    }),
+  ],
+
+  treeshake,
+} satisfies RollupOptions;
+
+const flatCJS = {
+  input: "src/flat.ts",
+
+  output: [
     {
       file: pkg.exports["./flat"].require,
       format: "cjs",
@@ -62,7 +99,18 @@ const flat = {
     rollupPluginAutoExternal(),
     rollupPluginTs({
       transpileOnly: true,
-      tsconfig: "tsconfig.build.json",
+      tsconfig: {
+        fileName: "tsconfig.build.json",
+        hook: (resolvedConfig) => ({
+          ...resolvedConfig,
+          paths: {
+            ...resolvedConfig.paths,
+            "#eslint-plugin-functional/conditional-imports/*": [
+              "src/utils/conditional-imports/cjs/*",
+            ],
+          },
+        }),
+      },
     }),
     rollupPluginDeassert({
       include: ["**/*.{js,ts}"],
@@ -72,4 +120,40 @@ const flat = {
   treeshake,
 } satisfies RollupOptions;
 
-export default [classic, flat];
+const flatESM = {
+  input: "src/flat.ts",
+
+  output: [
+    {
+      file: pkg.exports["./flat"].import,
+      format: "esm",
+      sourcemap: false,
+    },
+  ],
+
+  plugins: [
+    rollupPluginAutoExternal(),
+    rollupPluginTs({
+      transpileOnly: true,
+      tsconfig: {
+        fileName: "tsconfig.build.json",
+        hook: (resolvedConfig) => ({
+          ...resolvedConfig,
+          paths: {
+            ...resolvedConfig.paths,
+            "#eslint-plugin-functional/conditional-imports/*": [
+              "src/utils/conditional-imports/esm/*",
+            ],
+          },
+        }),
+      },
+    }),
+    rollupPluginDeassert({
+      include: ["**/*.{js,ts}"],
+    }),
+  ],
+
+  treeshake,
+} satisfies RollupOptions;
+
+export default [classicCJS, classicESM, flatCJS, flatESM];