+++ src/utils/conditional-imports/esm/ts-api-utils.ts
@@ -1,0 +1,10 @@
+import type tsApiUtils from "ts-api-utils";
+
+import ts from "#eslint-plugin-functional/conditional-imports/typescript";
+
+export default await ((): Promise<typeof tsApiUtils | undefined> => {
+  if (ts !== undefined) {
+    return import("ts-api-utils").catch(() => undefined);
+  }
+  return Promise.resolve(undefined);
+})();