+++ src/utils/conditional-imports/esm/typescript.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line unicorn/import-style
+export default await import("typescript")
+  .then((module) => module.default)
+  .catch(() => undefined);